requirements.txt
environs==9.5.0
python-telegram-bot==13.15
requests==2.28.2
